requirements-dev.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0

# Test data generation
faker>=19.3.1

# API testing
httpx>=0.24.1

# FastAPI and dependencies
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.7
databases>=0.8.0

# Utilities
python-dotenv>=1.0.0

